requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.15; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
appnope==0.1.4; python_version >= '3.6'
asttokens==3.0.0; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.3; python_version >= '3.7'
comm==0.2.3; python_version >= '3.8'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
debugpy==1.8.16; python_version >= '3.8'
decorator==5.2.1; python_version >= '3.8'
executing==2.2.0; python_version >= '3.8'
filelock==3.19.1; python_version >= '3.9'
frozenlist==1.7.0; python_version >= '3.9'
fsspec==2025.7.0; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
hf-xet==1.1.8; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.1; python_version >= '3.9'
huggingface-hub==0.34.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
iniconfig==2.1.0; python_version >= '3.8'
ipykernel==6.30.1; python_version >= '3.9'
ipython==9.4.0; python_version >= '3.11'
ipython-pygments-lexers==1.1.1; python_version >= '3.8'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.5.1; python_version >= '3.9'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jupyter-client==8.6.3; python_version >= '3.8'
jupyter-core==5.8.1; python_version >= '3.8'
langchain==0.3.27; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.27; python_version >= '3.9'
langchain-core==0.3.74; python_version >= '3.9'
langchain-huggingface==0.3.1; python_version >= '3.9'
langchain-ollama==0.3.7; python_version >= '3.9'
langchain-text-splitters==0.3.9; python_version >= '3.9'
langgraph==0.6.6; python_version >= '3.9'
langgraph-checkpoint==2.1.1; python_version >= '3.9'
langgraph-prebuilt==0.6.4; python_version >= '3.9'
langgraph-sdk==0.2.3; python_version >= '3.9'
langsmith==0.4.16; python_version >= '3.9'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
matplotlib-inline==0.1.7; python_version >= '3.8'
mpmath==1.3.0
multidict==6.6.4; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.5; python_version >= '3.11'
numpy==2.3.2; python_version >= '3.11'
ollama==0.5.3; python_version >= '3.8'
orjson==3.11.2; python_version >= '3.9'
ormsgpack==1.10.0; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
parso==0.8.5; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==11.3.0; python_version >= '3.9'
platformdirs==4.3.8; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
prompt-toolkit==3.0.51; python_version >= '3.8'
propcache==0.3.2; python_version >= '3.9'
psutil==7.0.0; python_version >= '3.6'
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.10.1; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pypdf==6.0.0; python_version >= '3.9'
pytest==8.4.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
pyzmq==27.0.2; python_version >= '3.8'
regex==2025.7.34; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
safetensors==0.6.2; python_version >= '3.9'
scikit-learn==1.7.1; python_version >= '3.10'
scipy==1.16.1; python_version >= '3.11'
sentence-transformers==5.1.0; python_version >= '3.9'
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.43; python_version >= '3.7'
stack-data==0.6.3
sympy==1.14.0; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tokenizers==0.21.4; python_version >= '3.9'
torch==2.8.0; python_full_version >= '3.9.0'
tornado==6.5.2; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
transformers==4.55.4; python_full_version >= '3.9.0'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspect==0.9.0
typing-inspection==0.4.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
wcwidth==0.2.13
xxhash==3.5.0; python_version >= '3.7'
yarl==1.20.1; python_version >= '3.9'
zstandard==0.24.0; python_version >= '3.9'
